+++ src/Sylius/Behat/Page/Shop/Taxon/ShowPage.php
@@ -16,8 +16,31 @@ use Sylius\Behat\Page\SymfonyPage;
 /**
  * @author Anna Walasek <anna.walasek@lakion.com>
  */
-class ShowPage extends SymfonyPage
+class ShowPage extends SymfonyPage implements ShowPageInterface
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function isProductInList($productName)
+    {
+        $elements= $this->getDocument()->findAll('css', 'div.header');
+        foreach ($elements as $element) {
+            if ($productName === $element->getText()) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isEmpty()
+    {
+        return false !== strpos($this->getDocument()->find('css', '.message')->getText(), 'There are no products to display');
+    }
+    
     /**
      * {@inheritdoc}
      */
