+++ src/Sylius/Behat/Resources/config/services/pages/shop/product.xml
@@ -14,11 +14,9 @@
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <parameters>
         <parameter key="sylius.behat.page.shop.product.show.class">Sylius\Behat\Page\Shop\Product\ShowPage</parameter>
-        <parameter key="sylius.behat.page.shop.product.index.class">Sylius\Behat\Page\Shop\Product\IndexPage</parameter>
     </parameters>
 
     <services>
         <service id="sylius.behat.page.shop.product.show" class="%sylius.behat.page.shop.product.show.class%" parent="sylius.behat.symfony_page" scope="scenario" public="false" />
-        <service id="sylius.behat.page.shop.product.index" class="%sylius.behat.page.shop.product.index.class%" parent="sylius.behat.symfony_page" scope="scenario" public="false" />
     </services>
 </container>
