+++ src/Sylius/Behat/Resources/config/suites/ui/product/viewing_products.yml
@@ -18,6 +18,7 @@ default:
                 - sylius.behat.context.setup.currency
                 - sylius.behat.context.setup.taxonomy
 
+                - sylius.behat.context.ui.shop.current_channel
                 - sylius.behat.context.ui.shop.product
             filters:
                 tags: "@viewing_products && @ui"
