+++ src/Sylius/Bundle/ShopBundle/Resources/config/routing/partial.yml
@@ -1,0 +1,3 @@
+sylius_shop_partial_product:
+    resource: "@SyliusShopBundle/Resources/config/routing/partial/product.yml"
+    prefix: /products
